@@ -14,7 +14,7 @@ import 'package:stream_chat/stream_chat.dart';
 class RetryQueue {
   /// Instantiate a new RetryQueue object
   RetryQueue({
-    @required this.channel,
+    required this.channel,
     this.logger,
   }) {
     _retryPolicy = channel.client.retryPolicy;
@@ -28,29 +28,29 @@ class RetryQueue {
   final Channel channel;
 
   /// The logger associated to this queue
-  final Logger logger;
+  final Logger? logger;
 
   final _subscriptions = <StreamSubscription>[];
 
   void _listenConnectionRecovered() {
     _subscriptions
         .add(channel.client.on(EventType.connectionRecovered).listen((event) {
-      if (!_isRetrying && event.online) {
+      if (!_isRetrying && event.online!) {
         _startRetrying();
       }
     }));
   }
 
-  final HeapPriorityQueue<Message> _messageQueue = HeapPriorityQueue(_byDate);
+  final HeapPriorityQueue<Message?> _messageQueue = HeapPriorityQueue(_byDate);
   bool _isRetrying = false;
-  RetryPolicy _retryPolicy;
+  RetryPolicy? _retryPolicy;
 
   /// Add a list of messages
-  void add(List<Message> messages) {
+  void add(List<Message?> messages) {
     logger?.info('added ${messages.length} messages');
     final messageList = _messageQueue.toList();
     _messageQueue.addAll(messages
-        .where((element) => !messageList.any((m) => m.id == element.id)));
+        .where((element) => !messageList.any((m) => m!.id == element!.id)));
 
     if (_messageQueue.isNotEmpty && !_isRetrying) {
       _startRetrying();
@@ -60,10 +60,10 @@ class RetryQueue {
   Future<void> _startRetrying() async {
     logger?.info('start retrying');
     _isRetrying = true;
-    final retryPolicy = _retryPolicy.copyWith(attempt: 0);
+    final retryPolicy = _retryPolicy!.copyWith(attempt: 0);
 
     while (_messageQueue.isNotEmpty) {
-      final message = _messageQueue.first;
+      final message = _messageQueue.first!;
       try {
         logger?.info('retry attempt ${retryPolicy.attempt}');
         await _sendMessage(message);
@@ -72,7 +72,7 @@ class RetryQueue {
         logger?.info('now ${_messageQueue.length} messages in the queue');
         retryPolicy.attempt = 0;
       } catch (error) {
-        ApiError apiError;
+        ApiError? apiError;
         if (error is DioError) {
           if (error.type == DioErrorType.response) {
             _messageQueue.remove(message);
@@ -84,7 +84,7 @@ class RetryQueue {
           );
         } else if (error is ApiError) {
           apiError = error;
-          if (apiError.status?.toString()?.startsWith('4') == true) {
+          if (apiError.status?.toString().startsWith('4') == true) {
             _messageQueue.remove(message);
             return;
           }
@@ -101,6 +101,7 @@ class RetryQueue {
         }
 
         retryPolicy.attempt++;
+
         final timeout = retryPolicy.retryTimeout(
           channel.client,
           retryPolicy.attempt,
@@ -112,13 +113,13 @@ class RetryQueue {
     _isRetrying = false;
   }
 
-  void _sendFailedEvent(Message message) {
-    final newStatus = message.status == MessageSendingStatus.sending
+  void _sendFailedEvent(Message? message) {
+    final newStatus = message!.status == MessageSendingStatus.sending
         ? MessageSendingStatus.failed
         : (message.status == MessageSendingStatus.updating
             ? MessageSendingStatus.failed_update
             : MessageSendingStatus.failed_delete);
-    channel.state.addMessage(message.copyWith(
+    channel.state!.addMessage(message.copyWith(
       status: newStatus,
     ));
   }
@@ -141,20 +142,20 @@ class RetryQueue {
       final messageList = _messageQueue.toList();
       if (event.message != null) {
         final messageIndex =
-            messageList.indexWhere((m) => m.id == event.message.id);
+            messageList.indexWhere((m) => m!.id == event.message!.id);
         if (messageIndex == -1 &&
             [
               MessageSendingStatus.failed_update,
               MessageSendingStatus.failed,
               MessageSendingStatus.failed_delete,
-            ].contains(event.message.status)) {
+            ].contains(event.message!.status)) {
           logger?.info('add message from events');
           add([event.message]);
         } else if (messageIndex != -1 &&
             [
               MessageSendingStatus.sent,
               null,
-            ].contains(event.message.status)) {
+            ].contains(event.message!.status)) {
           _messageQueue.remove(messageList[messageIndex]);
         }
       }
@@ -167,14 +168,14 @@ class RetryQueue {
     _subscriptions.forEach((s) => s.cancel());
   }
 
-  static int _byDate(Message m1, Message m2) {
-    final date1 = _getMessageDate(m1);
-    final date2 = _getMessageDate(m2);
+  static int _byDate(Message? m1, Message? m2) {
+    final date1 = _getMessageDate(m1!)!;
+    final date2 = _getMessageDate(m2!)!;
 
     return date1.compareTo(date2);
   }
 
-  static DateTime _getMessageDate(Message m1) {
+  static DateTime? _getMessageDate(Message m1) {
     switch (m1.status) {
       case MessageSendingStatus.failed_delete:
       case MessageSendingStatus.deleting:
