@@ -17,13 +17,13 @@ class Command {
       _$CommandFromJson(json);
 
   /// The name of the command
-  final String name;
+  final String? name;
 
   /// The description explaining the command
-  final String description;
+  final String? description;
 
   /// The arguments of the command
-  final String args;
+  final String? args;
 
   /// Serialize to json
   Map<String, dynamic> toJson() => _$CommandToJson(this);
