@@ -44,7 +44,7 @@ enum MessageSendingStatus {
 class Message {
   /// Constructor used for json serialization
   Message({
-    String id,
+    String? id,
     this.text,
     this.type,
     this.attachments,
@@ -67,7 +67,7 @@ class Message {
     this.user,
     this.pinned = false,
     this.pinnedAt,
-    DateTime pinExpires,
+    DateTime? pinExpires,
     this.pinnedBy,
     this.extraData,
     this.deletedAt,
@@ -77,15 +77,15 @@ class Message {
         pinExpires = pinExpires?.toUtc();
 
   /// Create a new instance from a json
-  factory Message.fromJson(Map<String, dynamic> json) => _$MessageFromJson(
-      Serialization.moveToExtraDataFromRoot(json, topLevelFields));
+  factory Message.fromJson(Map<String, dynamic>? json) => _$MessageFromJson(
+      Serialization.moveToExtraDataFromRoot(json, topLevelFields)!);
 
   /// The message ID. This is either created by Stream or set client side when
   /// the message is added.
   final String id;
 
   /// The text of this message
-  final String text;
+  final String? text;
 
   /// The status of a sending message
   @JsonKey(ignore: true)
@@ -93,99 +93,99 @@ class Message {
 
   /// The message type
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final String type;
+  final String? type;
 
   /// The list of attachments, either provided by the user or generated from a
   /// command or as a result of URL scraping.
   @JsonKey(includeIfNull: false)
-  final List<Attachment> attachments;
+  final List<Attachment>? attachments;
 
   /// The list of user mentioned in the message
   @JsonKey(toJson: Serialization.userIds)
-  final List<User> mentionedUsers;
+  final List<User>? mentionedUsers;
 
   /// A map describing the count of number of every reaction
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final Map<String, int> reactionCounts;
+  final Map<String?, int>? reactionCounts;
 
   /// A map describing the count of score of every reaction
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final Map<String, int> reactionScores;
+  final Map<String?, int>? reactionScores;
 
   /// The latest reactions to the message created by any user.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final List<Reaction> latestReactions;
+  final List<Reaction>? latestReactions;
 
   /// The reactions added to the message by the current user.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final List<Reaction> ownReactions;
+  final List<Reaction>? ownReactions;
 
   /// The ID of the parent message, if the message is a thread reply.
-  final String parentId;
+  final String? parentId;
 
   /// A quoted reply message
   @JsonKey(toJson: Serialization.readOnly)
-  final Message quotedMessage;
+  final Message? quotedMessage;
 
   /// The ID of the quoted message, if the message is a quoted reply.
-  final String quotedMessageId;
+  final String? quotedMessageId;
 
   /// Reserved field indicating the number of replies for this message.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final int replyCount;
+  final int? replyCount;
 
   /// Reserved field indicating the thread participants for this message.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final List<User> threadParticipants;
+  final List<User>? threadParticipants;
 
   /// Check if this message needs to show in the channel.
-  final bool showInChannel;
+  final bool? showInChannel;
 
   /// If true the message is silent
-  final bool silent;
+  final bool? silent;
 
   /// If true the message will not send a push notification
-  final bool skipPush;
+  final bool? skipPush;
 
   /// If true the message is shadowed
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final bool shadowed;
+  final bool? shadowed;
 
   /// A used command name.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final String command;
+  final String? command;
 
   /// Reserved field indicating when the message was created.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final DateTime createdAt;
+  final DateTime? createdAt;
 
   /// Reserved field indicating when the message was updated last time.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final DateTime updatedAt;
+  final DateTime? updatedAt;
 
   /// User who sent the message
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final User user;
+  final User? user;
 
   /// If true the message is pinned
-  final bool pinned;
+  final bool? pinned;
 
   /// Reserved field indicating when the message was pinned
   @JsonKey(toJson: Serialization.readOnly)
-  final DateTime pinnedAt;
+  final DateTime? pinnedAt;
 
   /// Reserved field indicating when the message will expire
   ///
   /// if `null` message has no expiry
-  final DateTime pinExpires;
+  final DateTime? pinExpires;
 
   /// Reserved field indicating who pinned the message
   @JsonKey(toJson: Serialization.readOnly)
-  final User pinnedBy;
+  final User? pinnedBy;
 
   /// Message custom extraData
   @JsonKey(includeIfNull: false)
-  final Map<String, dynamic> extraData;
+  final Map<String, dynamic>? extraData;
 
   /// True if the message is a system info
   bool get isSystem => type == 'system';
@@ -198,7 +198,7 @@ class Message {
 
   /// Reserved field indicating when the message was deleted.
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final DateTime deletedAt;
+  final DateTime? deletedAt;
 
   /// Known top level fields.
   /// Useful for [Serialization] methods.
@@ -239,35 +239,35 @@ class Message {
 
   /// Creates a copy of [Message] with specified attributes overridden.
   Message copyWith({
-    String id,
-    String text,
-    String type,
-    List<Attachment> attachments,
-    List<User> mentionedUsers,
-    Map<String, int> reactionCounts,
-    Map<String, int> reactionScores,
-    List<Reaction> latestReactions,
-    List<Reaction> ownReactions,
-    String parentId,
-    Message quotedMessage,
-    String quotedMessageId,
-    int replyCount,
-    List<User> threadParticipants,
-    bool showInChannel,
-    bool shadowed,
-    bool silent,
-    String command,
-    DateTime createdAt,
-    DateTime updatedAt,
-    DateTime deletedAt,
-    User user,
-    bool pinned,
-    DateTime pinnedAt,
-    Object pinExpires = _pinExpires,
-    User pinnedBy,
-    Map<String, dynamic> extraData,
-    MessageSendingStatus status,
-    bool skipPush,
+    String? id,
+    String? text,
+    String? type,
+    List<Attachment>? attachments,
+    List<User>? mentionedUsers,
+    Map<String?, int>? reactionCounts,
+    Map<String?, int>? reactionScores,
+    List<Reaction>? latestReactions,
+    List<Reaction>? ownReactions,
+    String? parentId,
+    Message? quotedMessage,
+    String? quotedMessageId,
+    int? replyCount,
+    List<User>? threadParticipants,
+    bool? showInChannel,
+    bool? shadowed,
+    bool? silent,
+    String? command,
+    DateTime? createdAt,
+    DateTime? updatedAt,
+    DateTime? deletedAt,
+    User? user,
+    bool? pinned,
+    DateTime? pinnedAt,
+    Object? pinExpires = _pinExpires,
+    User? pinnedBy,
+    Map<String, dynamic>? extraData,
+    MessageSendingStatus? status,
+    bool? skipPush,
   }) {
     assert(() {
       if (pinExpires is! DateTime &&
@@ -305,7 +305,8 @@ class Message {
       pinned: pinned ?? this.pinned,
       pinnedAt: pinnedAt ?? this.pinnedAt,
       pinnedBy: pinnedBy ?? this.pinnedBy,
-      pinExpires: pinExpires == _pinExpires ? this.pinExpires : pinExpires,
+      pinExpires:
+          pinExpires == _pinExpires ? this.pinExpires : pinExpires as DateTime?,
       skipPush: skipPush ?? this.skipPush,
     );
   }
@@ -355,13 +356,13 @@ class TranslatedMessage extends Message {
   TranslatedMessage(this.i18n);
 
   /// Create a new instance from a json
-  factory TranslatedMessage.fromJson(Map<String, dynamic> json) =>
+  factory TranslatedMessage.fromJson(Map<String, dynamic>? json) =>
       _$TranslatedMessageFromJson(
-        Serialization.moveToExtraDataFromRoot(json, topLevelFields),
+        Serialization.moveToExtraDataFromRoot(json, topLevelFields)!,
       );
 
   /// A Map of
-  final Map<String, String> i18n;
+  final Map<String, String>? i18n;
 
   /// Known top level fields.
   /// Useful for [Serialization] methods.
