@@ -49,7 +49,7 @@ void main() {
         'https://media0.giphy.com/media/3o7TKnCdBx5cMg0qti/giphy.gif',
       );
       expect(attachment.actions, hasLength(3));
-      expect(attachment.actions[0], isA<Action>());
+      expect(attachment.actions![0], isA<Action>());
     });
 
     test('should serialize to json correctly', () {
